Remove useMessagesSelection
